requirements.txt
websockets==12.0
msgpack==1.0.8
python-dotenv==1.0.1
colorama==0.4.6
aiohttp==3.9.5
